Shifted consume function to func/cinsume/consume.cpp
