+++ Parser.cpp
@@ -26,20 +26,6 @@ ASTNode Parser::parse()
     return root;
 };
 
-Token Parser::consume(const string &expectedType)
-{
-    Token token = peek();
-    if (token.type == expectedType)
-    {
-        currentIndex++;
-        return token;
-    }
-    else
-    {
-        throw runtime_error("Unexpected token: expected " + expectedType + ", got " + token.type);
-    }
-}
-
 ASTNode Parser::parseAssignment()
 {
     Token identifier = consume("IDENTIFIER");
